Replaced the regex implementatino with hyperscan which supports streaming
Streaming is required for reducing memory usage while still granting
the same search results, by streaming small 64kB blocks to the scanner
instead of one complete memory mapping.

Hyperscan is also faster, so win win. But it's not pure rust and
requirse an additional library to be installed while compiling. The
library itself is linked statically to the binary, so end-user does
not need it.

Accompanies the Hyperscan license due to said statically linking in
releases.